Allow to set some overrides for the in-cluster gitea
The VP operator reads the confimap to get some defaults for its inner
settings. Expose these gitea values so it is much easier to test new
gitea charts.

Tested this with a pattern setting the following in values-global.yaml:

    main:
      clusterGroupName: datacenter
      multiSourceConfig:
        enabled: true
        clusterGroupChartVersion: 0.9.*
      git:
        repoUpstreamURL: https://github.com/mbaldessari/industrial-edge
      gitea:
        chartVersion: 0.0.3
        chartName: gitea
        helmRepoUrl: https://acksyn.org/files/helm-repo

And correctly get the in-cluster-gitea argo application pointing to the
new repository.